backend/requirements.txt
# Core Framework
Flask>=2.0.0
Werkzeug>=2.0.0

# AI and Machine Learning
google-generativeai>=0.3.0
tensorflow>=2.8.0
Pillow>=8.3.0
numpy>=1.21.0

# Database
pymongo>=4.0.0

# Security and Authentication
flask-session>=0.4.0
flask-wtf>=1.0.0
PyJWT>=2.4.0
cryptography>=3.4.8

# HTTP and Requests
requests>=2.26.0
urllib3>=1.26.0

# Monitoring and Logging
psutil>=5.8.0

# Development and Testing
pytest>=6.2.0
pytest-flask>=1.2.0
python-dotenv>=0.19.0

# Production Server
gunicorn>=20.1.0
gevent>=21.8.0

# File Processing
python-magic>=0.4.24

# Validation
marshmallow>=3.14.0
email-validator>=1.1.3
